refactor: combine product actions with SelectedVariant (#96)
